Changes for planner, joins, perf counters
Changed the way that the planner represents nested subqueries, so that
the outermost query is always at index 0, and the innermost query is
always the last query in the list of enclosing queries.  This matches
how other parts of NanoDB work, so it seemed wiser to do it this way.

Added semijoin and antijoin as join types to the enum.

Made a field private on the property registry, because it really didn't
need to be public...

Added a basic performance-counter service to NanoDB so that we can
instrument various parts of the database and see how it behaves.  The
perf-counter code should be completely thread-safe, which means it
probably actually isn't.