Changes to support statistics-collection in heap files
Implemented the HeapTupleFileManager.saveMetadata() method, which saves
the schema and statistics for a given tuple file.  This basically
involved refactoring some existing code in the manager, so it wasn't
complicated.

Added a sanity-check to the heap-file page tuple implementation to
verify that the page-offset provided to the constructor actually matches
the offset stored in the slot that is also specified to the constructor.