A simple performance-counter framework.
Added a new performance-counter framework to NanoDB so that we can
measure the performance of various parts of the database in an absolute,
repeatable way.  Right now, individual performance counters can be
accessed via the PERF_COUNTER('...') function, and they can be cleared
via the CLEAR_PERF_COUNTER('...') function.

Finally, a few example performance counters have been added to the File
Manager to count block reads, writes and disk seeks.