Fixed a bug in buffer manager's page-pinning code.
The mapping that tracks pinned pages was only using the page number to
identify pages, which was a stupid error.  It needs to also include the
file that the page is from.

The updated version uses a special class that includes both the file and
the page number, so that we get type-safety in our mapping, and the
compiler can prevent us from making some stupid bugs.